requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0.1
django-material==1.2.2
django-widget-tweaks==1.4.1
gunicorn==19.7.1
psycopg2==2.7.3.2
python-decouple==3.1
pytz==2017.3
static3==0.7.0
Unipath==1.1
whitenoise==3.3.1
